feat: first review pass
