+++ packages/swapper/src/swappers/RelaySwapper/getTradeRate/getTradeRate.ts
@@ -2,25 +2,33 @@ import type { ChainId } from '@shapeshiftoss/caip'
 import type { Result } from '@sniptt/monads'
 
 import type {
-  CommonTradeQuoteInput,
   GetTradeRateInput,
   SwapErrorRight,
   SwapperDeps,
+  TradeRate,
   TradeRateStep,
 } from '../../../types'
-import { getTrade } from '../getTradeQuote/getTradeQuote'
-import type { RelayTradeRate } from '../utils/types'
+import { getRelayTradeRate } from '../utils/getTrade'
 
 export const getTradeRate = async (
-  input: CommonTradeQuoteInput | GetTradeRateInput,
+  input: GetTradeRateInput,
   deps: SwapperDeps,
   relayChainMap: Record<ChainId, number>,
-): Promise<Result<RelayTradeRate[], SwapErrorRight>> => {
-  const ratesResult = await getTrade({
-    input,
-    deps,
-    relayChainMap,
-  })
+): Promise<Result<TradeRate[], SwapErrorRight>> => {
+  const args = {
+    quoteOrRate: 'rate' as const,
+    buyAsset: input.buyAsset,
+    receiveAddress: input.receiveAddress,
+    sellAmountIncludingProtocolFeesCryptoBaseUnit:
+      input.sellAmountIncludingProtocolFeesCryptoBaseUnit,
+    sellAsset: input.sellAsset,
+    sendAddress: input.sendAddress,
+    accountNumber: input.accountNumber,
+    affiliateBps: input.affiliateBps,
+    potentialAffiliateBps: input.potentialAffiliateBps,
+  }
+
+  const ratesResult = await getRelayTradeRate(args, deps, relayChainMap)
 
   return ratesResult.map(rates =>
     rates.map(rate => ({
